LayoutTests/imported/w3c/web-platform-tests/xhr/setrequestheader-allow-empty-value-expected.txt


FAIL XMLHttpRequest: setRequestHeader() - empty header () assert_equals: expected "X-Empty: \n" but got "X-Empty:  \n"
PASS XMLHttpRequest: setRequestHeader() - empty header (null) 
PASS XMLHttpRequest: setRequestHeader() - empty header (undefined) 
